refactor: ContextMenu.plugin
