+++ webapp/javascript/components/TimelineChart/AnnotationMark/styles.module.scss
@@ -19,5 +19,5 @@
 }
 
 .form {
-  width: 160px;
+  width: 180px;
 }
