+++ webapp/javascript/components/TimelineChart/TimelineChart.tsx
@@ -9,6 +9,7 @@ import 'react-flot/flot/jquery.flot.crosshair.min';
 import './TimelineChartPlugin';
 import './Tooltip.plugin';
 import './Annotations.plugin';
+import './ContextMenu.plugin';
 
 interface TimelineChartProps {
   onSelect: (from: string, until: string) => void;
